2012-10-02 12:55 UTC+0700 Andi Jahja <xharbour/AT/telkom/net/id>   * contrib/filemem/filemem.c   * contrib/gtwvg/gtwvg.c   * contrib/gtwvw/gtwvw.c   * contrib/hbbtree/hb_btree.h   * contrib/libnf/dispc.c   * contrib/misc/gauge.c   * contrib/misc/hb_f.c   * contrib/misc/num.c   * contrib/misc/stack.c   * contrib/misc/strfmt.c   * contrib/misc/stringsx.c   * contrib/pdflite/pdflib/p_document.c   * contrib/pdflite/tiff/tif_unix.c   * contrib/rddsql/hbrddsql.h   * contrib/rddsql/sqlbase.c   * contrib/rddsql/sqlmix.c   * contrib/sddsqlt3/sddsqlt3.c   * contrib/what32/source/_wincdlg.c   * contrib/what32/source/_windlg.c   * contrib/what32/source/_winmain.c   * contrib/what32/source/_winsys.c   * include/arc4.h   * include/classes.h   * include/ctstr.h   * include/hashapi.h   * include/hbapi.h   * include/hbapicdp.h   * include/hbapidbg.h   * include/hbapifs.h   * include/hbapigt.h   * include/hbapiitm.h   * include/hbapistr.h   * include/hbcomp.h   * include/hbcompat.h   * include/hbdate.h   * include/hbdefs.h   * include/hbexprb.c   * include/hbexprc.c   * include/hbexprop.h   * include/hbfast.h   * include/hbgtcore.h   * include/hbhash.h   * include/hbjson.h   * include/hbmacro.h   * include/hbnxs.h   * include/hbpp.h   * include/hbrddfpt.h   * include/hbstack.h   * include/hbvm.h   * include/hbvmpub.h   * include/hbxml.h   * include/simplex.c   * include/thread.h   * source/common/expropt1.c   * source/common/expropt2.c   * source/common/hbarch.c   * source/common/hbdate.c   * source/common/hbfopen.c   * source/common/hbfsapi.c   * source/common/hbgete.c   * source/common/hbhash.c   * source/common/hbmem.c   * source/common/hbprintf.c   * source/common/hbstr.c   * source/common/hbtrace.c   * source/common/hbver.c   * source/common/hbwince.c   * source/common/reserved.c   * source/compiler/calconst.c   * source/compiler/cmdcheck.c   * source/compiler/comptool.c   * source/compiler/expropta.c   * source/compiler/exproptb.c   * source/compiler/exproptc.c   * source/compiler/genc.c   * source/compiler/gencc.c   * source/compiler/gencc1.c   * source/compiler/gencobj.c   * source/compiler/genhrb.c   * source/compiler/harbour.c   * source/compiler/harbour.slx   * source/compiler/harbour.sly   * source/compiler/harbouryy.c   * source/compiler/hbdbginf.c   * source/compiler/hbdead.c   * source/compiler/hbfix.c   * source/compiler/hbfunchk.c   * source/compiler/hbgenerr.c   * source/compiler/hbident.c   * source/compiler/hblbl.c   * source/compiler/hbpcode.c   * source/compiler/hbpcstat.c   * source/compiler/hbslex.c   * source/compiler/hbstripl.c   * source/compiler/hbstrong.c   * source/compiler/hbusage.c   * source/compiler/ppcomp.c   * source/ct/addascii.c   * source/ct/asciisum.c   * source/ct/ascpos.c   * source/ct/atadjust.c   * source/ct/atnum.c   * source/ct/atrepl.c   * source/ct/blank.c   * source/ct/charevod.c   * source/ct/charlist.c   * source/ct/charmix.c   * source/ct/charop.c   * source/ct/charrepl.c   * source/ct/count.c   * source/ct/ctc.c   * source/ct/ctchksum.c   * source/ct/ctcrypt.c   * source/ct/ctstr.c   * source/ct/ctwin.c   * source/ct/dattime2.c   * source/ct/dattime3.c   * source/ct/disk.c   * source/ct/envparam.c   * source/ct/intneg.c   * source/ct/justify.c   * source/ct/maxline.c   * source/ct/misc2.c   * source/ct/numat.c   * source/ct/numline.c   * source/ct/pack.c   * source/ct/pos1.c   * source/ct/pos2.c   * source/ct/posdiff.c   * source/ct/relation.c   * source/ct/screen1.c   * source/ct/screen2.c   * source/ct/strfile.c   * source/ct/tab.c   * source/ct/token1.c   * source/ct/token2.c   * source/ct/wordrepl.c   * source/ct/wordtoch.c   * source/macro/macro.slx   * source/macro/macro.y   * source/macro/macroa.c   * source/macro/macrob.c   * source/macro/macroc.c   * source/macro/macroslx.c   * source/macro/macroyy.c   * source/odbc/odbc.c   * source/pp/ppcore.c   * source/pp/ppgen.c   * source/pp/pplib.c   * source/pp/pplib3.c   * source/rdd/bmdbfcdx/bmdbfcdx1.c   * source/rdd/bmdbfcdx/redbfcdx1.c   * source/rdd/dbcmd.c   * source/rdd/dbf1.c   * source/rdd/dbf1net.c   * source/rdd/dbfcdx/dbfcdx1.c   * source/rdd/dbffpt/dbffpt1.c   * source/rdd/dbffpt/redbffpt1.c   * source/rdd/dbfmdx/dbfmdx1.c   * source/rdd/dbfnsx/dbfnsx1.c   * source/rdd/dbfntx/dbfntx1.c   * source/rdd/delim1.c   * source/rdd/hbsix/sxcompr.c   * source/rdd/hbsix/sxcrypt.c   * source/rdd/hbsix/sxfname.c   * source/rdd/hbsix/sxtable.c   * source/rdd/hsx/hsx.c   * source/rdd/sdf1.c   * source/rdd/usrrdd/usrrdd.c   * source/rdd/wafunc.c   * source/rtl/ampm.c   * source/rtl/arc4.c   * source/rtl/at.c   * source/rtl/binnum.c   * source/rtl/binnumx.c   * source/rtl/bkgtsks.c   * source/rtl/cdpapi.c   * source/rtl/chruni.c   * source/rtl/console.c   * source/rtl/dateshb.c   * source/rtl/dbf2txt.c   * source/rtl/defpath.c   * source/rtl/descend.c   * source/rtl/direct.c   * source/rtl/dllcall.c   * source/rtl/file.c   * source/rtl/filebuf.c   * source/rtl/filenet.c   * source/rtl/filesys.c   * source/rtl/fparse.c   * source/rtl/fstemp.c   * source/rtl/gete.c   * source/rtl/gtapi.c   * source/rtl/gtcgi/gtcgi.c   * source/rtl/gtclip.c   * source/rtl/gtgui/gtgui.c   * source/rtl/gtpca/gtpca.c   * source/rtl/gtstd/gtstd.c   * source/rtl/gtwin/gtwin.c   * source/rtl/gtwvt/gtwvt.c   * source/rtl/hardcr.c   * source/rtl/hbbffnc.c   * source/rtl/hbchksum.c   * source/rtl/hbcomprs.c   * source/rtl/hbcrc32.c   * source/rtl/hbcrypt.c   * source/rtl/hbffind.c   * source/rtl/hbgtcore.c   * source/rtl/hbhex2n.c   * source/rtl/hbip.c   * source/rtl/hbjson.c   * source/rtl/hbmd5.c   * source/rtl/hboutdbg.c   * source/rtl/hbrand.c   * source/rtl/hbserv.c   * source/rtl/hbsha1.c   * source/rtl/hbsha1hm.c   * source/rtl/hbsha2.c   * source/rtl/hbsha2hm.c   * source/rtl/hbsocket.c   * source/rtl/hbsockhb.c   * source/rtl/hbsrlraw.c   * source/rtl/hbtoken.c   * source/rtl/hbwinole.c   * source/rtl/hbxml.c   * source/rtl/hbzlibgz.c   * source/rtl/inkey.c   * source/rtl/inkeyapi.c   * source/rtl/left.c   * source/rtl/len.c   * source/rtl/lennum.c   * source/rtl/memofile.c   * source/rtl/mlcfunc.c   * source/rtl/mtran.c   * source/rtl/oemansi.c   * source/rtl/padc.c   * source/rtl/padl.c   * source/rtl/padr.c   * source/rtl/philes.c   * source/rtl/rat.c   * source/rtl/readline.c   * source/rtl/regex.c   * source/rtl/replic.c   * source/rtl/right.c   * source/rtl/samples.c   * source/rtl/saverest.c   * source/rtl/seconds.c   * source/rtl/setkey.c   * source/rtl/soundex.c   * source/rtl/strcase.c   * source/rtl/strdel.c   * source/rtl/strings.c   * source/rtl/strmatch.c   * source/rtl/strpeek.c   * source/rtl/strtran.c   * source/rtl/stuff.c   * source/rtl/transfrm.c   * source/rtl/trim.c   * source/rtl/txtline.c   * source/rtl/valtostr.c   * source/rtl/version.c   * source/rtl/win32prn.prg   * source/rtl/winreg.prg   * source/rtl/xsavescr.c   * source/tip/base64x.c   * source/tip/encb64.prg   * source/tip/encqp.prg   * source/tip/thtml.prg   * source/tip/utils.c   * source/vm/arrays.c   * source/vm/arrayshb.c   * source/vm/asort.c   * source/vm/classes.c   * source/vm/cmdarg.c   * source/vm/codebloc.c   * source/vm/dbgentry.c   * source/vm/debug.c   * source/vm/dynsym.c   * source/vm/estack.c   * source/vm/eval.c   * source/vm/extend.c   * source/vm/fastitem.c   * source/vm/fm.c   * source/vm/garbage.c   * source/vm/hash.c   * source/vm/hbi18n.c   * source/vm/hvm.c   * source/vm/inet.c   * source/vm/itemapi.c   * source/vm/macro.c   * source/vm/memvars.c   * source/vm/proc.c   * source/vm/runner.c   * source/vm/set.c   * source/vm/strapi.c   * source/vm/thread.c   * utils/hbdoc/pdfhbdoc.c   * utils/hbpp/hbpp.c   * utils/hbpp/hbppcalc.c   * utils/hbpp/hbppcomp.c   * utils/hbpp/hbppcore.c   * utils/hbpp/hbppdef.h   * utils/hbpp/hbpptbl.c   * utils/hbpp/pragma.c   * utils/xbscript/xbs_harb.ch     ! First patches of HB_SIZE data type implementation (idea from Harbour)

+++ xharbour/contrib/filemem/filemem.c
@@ -205,7 +205,7 @@ static PHB_MEMFS_INODE memfsInodeAlloc( const char* szName )
    pInode->llSize = 0;
    pInode->llAlloc = HB_MEMFS_INITSIZE;
    pInode->pData = ( char * ) hb_xgrab( ( ULONG ) pInode->llAlloc );
-   memset( pInode->pData, 0, pInode->llAlloc );
+   memset( pInode->pData, 0, (size_t) pInode->llAlloc );
    pInode->szName = hb_strdup( szName );
 
    pInode->uiCount = 1;
@@ -551,7 +551,7 @@ HB_MEMFS_EXPORT ULONG hb_memfsWriteAt( HB_FHANDLE hFile, const void * pBuff, ULO
          llNewAlloc = llOffset + ( HB_FOFFSET ) ulCount;
 
       pInode->pData = ( char * ) hb_xrealloc( pInode->pData, ( ULONG ) llNewAlloc );
-      memset( pInode->pData + ( ULONG ) pInode->llAlloc, 0, llNewAlloc - pInode->llAlloc );
+      memset( pInode->pData + ( ULONG ) pInode->llAlloc, 0, (size_t) ( llNewAlloc - pInode->llAlloc ) );
       pInode->llAlloc = llNewAlloc;
    }
    memcpy( pInode->pData + ( ULONG ) llOffset, pBuff, ulCount );
@@ -605,7 +605,7 @@ HB_MEMFS_EXPORT BOOL hb_memfsTruncAt( HB_FHANDLE hFile, HB_FOFFSET llOffset )
          llNewAlloc = llOffset;
 
       pInode->pData = ( char * ) hb_xrealloc( pInode->pData, ( ULONG ) llNewAlloc );
-      memset( pInode->pData + ( ULONG ) pInode->llAlloc, 0, llNewAlloc - pInode->llAlloc );
+      memset( pInode->pData + ( ULONG ) pInode->llAlloc, 0, (size_t) ( llNewAlloc - pInode->llAlloc ) );
       pInode->llAlloc = llNewAlloc;
    }
    else if( ( pInode->llAlloc >> 2 ) > ( llOffset > HB_MEMFS_INITSIZE ? llOffset : HB_MEMFS_INITSIZE ) )
@@ -614,7 +614,7 @@ HB_MEMFS_EXPORT BOOL hb_memfsTruncAt( HB_FHANDLE hFile, HB_FOFFSET llOffset )
       pInode->pData = ( char * ) hb_xrealloc( pInode->pData, ( ULONG ) pInode->llAlloc );
    }
 
-   memset( pInode->pData + ( ULONG ) llOffset, 0, pInode->llAlloc - llOffset );
+   memset( pInode->pData + ( ULONG ) llOffset, 0, (size_t) ( pInode->llAlloc - llOffset ) );
 
    pInode->llSize = llOffset;
    HB_MEMFSMT_UNLOCK
